refactor: infrastructrue/config 파일구조 변경
