+++ src/main/java/com/process/clash/application/record/service/RecordDayBoundaryService.java
@@ -5,7 +5,7 @@ import com.process.clash.application.record.port.out.RecordActivityNotifierPort;
 import com.process.clash.application.record.port.out.RecordSessionRepositoryPort;
 import com.process.clash.application.user.exp.service.StudyTimeExpGrantService;
 import com.process.clash.domain.record.entity.RecordSession;
-import com.process.clash.infrastructure.config.RecordProperties;
+import com.process.clash.infrastructure.config.record.RecordProperties;
 import jakarta.transaction.Transactional;
 import java.time.Instant;
 import java.time.LocalDate;
