+++ src/main/java/com/process/clash/application/record/v2/service/GetTodayRecordV2Service.java
@@ -9,7 +9,7 @@ import com.process.clash.application.record.v2.port.out.RecordSessionV2Repositor
 import com.process.clash.application.user.user.exception.exception.notfound.UserNotFoundException;
 import com.process.clash.application.user.user.port.out.UserRepositoryPort;
 import com.process.clash.domain.record.v2.entity.RecordSessionV2;
-import com.process.clash.infrastructure.config.RecordProperties;
+import com.process.clash.infrastructure.config.record.RecordProperties;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
