+++ src/test/java/com/process/clash/application/record/v2/service/GetActivityStatisticsV2ServiceTest.java
@@ -19,7 +19,7 @@ import com.process.clash.domain.record.enums.MonitoredApp;
 import com.process.clash.domain.user.user.entity.User;
 import com.process.clash.domain.user.user.enums.Role;
 import com.process.clash.domain.user.user.enums.UserStatus;
-import com.process.clash.infrastructure.config.RecordProperties;
+import com.process.clash.infrastructure.config.record.RecordProperties;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
