+++ src/test/java/com/process/clash/application/record/v2/service/GetAllSubjectsV2ServiceTest.java
@@ -14,7 +14,7 @@ import com.process.clash.domain.record.v2.entity.RecordSessionV2;
 import com.process.clash.domain.record.v2.entity.RecordSubjectV2;
 import com.process.clash.domain.record.v2.entity.RecordTaskV2;
 import com.process.clash.domain.record.v2.enums.RecordSessionTypeV2;
-import com.process.clash.infrastructure.config.RecordProperties;
+import com.process.clash.infrastructure.config.record.RecordProperties;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
