+++ src/test/java/com/process/clash/application/record/v2/service/GetAllTasksV2ServiceTest.java
@@ -11,7 +11,7 @@ import com.process.clash.application.record.v2.port.out.RecordTaskV2RepositoryPo
 import com.process.clash.domain.record.v2.entity.RecordSessionV2;
 import com.process.clash.domain.record.v2.entity.RecordTaskV2;
 import com.process.clash.domain.record.v2.enums.RecordSessionTypeV2;
-import com.process.clash.infrastructure.config.RecordProperties;
+import com.process.clash.infrastructure.config.record.RecordProperties;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
