+++ src/test/java/com/process/clash/application/record/v2/service/GetTodayRecordV2ServiceTest.java
@@ -20,7 +20,7 @@ import com.process.clash.domain.record.v2.enums.RecordSessionTypeV2;
 import com.process.clash.domain.user.user.entity.User;
 import com.process.clash.domain.user.user.enums.Role;
 import com.process.clash.domain.user.user.enums.UserStatus;
-import com.process.clash.infrastructure.config.RecordProperties;
+import com.process.clash.infrastructure.config.record.RecordProperties;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
